+++ configs/app_config.h
@@ -8,7 +8,7 @@
 
 #include "iotconnect.h"
 
-#define IOTCONNECT_CPID "your-cpidd"
+#define IOTCONNECT_CPID "your-cpid"
 #define IOTCONNECT_ENV  "your-environment"
 
 // Device Unique ID
@@ -20,16 +20,49 @@
 // PEM format certificate and private key, if using X509 type auth.
 // For example:
 #define IOTCONNECT_DEVICE_CERT \
-"-----BEGIN CERTIFICATE-----\n"\
-"...\n"\
+"-----BEGIN CERTIFICATE-----\n" \
+
 "-----END CERTIFICATE-----"
 
 #define IOTCONNECT_DEVICE_KEY \
-"-----BEGIN RSA PRIVATE KEY-----\n"\
-"...\n"\
+"-----BEGIN RSA PRIVATE KEY-----\n" \
+
 "-----END RSA PRIVATE KEY-----"
 
 // you can choose to use your own NTP server to obtain network time, or simply time.google.com for better stability
 #define IOTCONNECT_SNTP_SERVER "pool.ntp.org"
+
+//--------------------------------
+//-----------FOR OTA--------------
+//--------------------------------
+
+#define HTTP_SERVER_PORT    (443)
+
+/* Name of the JSON job file for HTTP  */
+#define OTA_HTTP_JOB_FILE    "dummy"
+
+#define ROOT_CA_CERTIFICATE \
+"-----BEGIN CERTIFICATE-----\n" \
+"MIIDjjCCAnagAwIBAgIQAzrx5qcRqaC7KGSxHQn65TANBgkqhkiG9w0BAQsFADBh\n" \
+"MQswCQYDVQQGEwJVUzEVMBMGA1UEChMMRGlnaUNlcnQgSW5jMRkwFwYDVQQLExB3\n" \
+"d3cuZGlnaWNlcnQuY29tMSAwHgYDVQQDExdEaWdpQ2VydCBHbG9iYWwgUm9vdCBH\n" \
+"MjAeFw0xMzA4MDExMjAwMDBaFw0zODAxMTUxMjAwMDBaMGExCzAJBgNVBAYTAlVT\n" \
+"MRUwEwYDVQQKEwxEaWdpQ2VydCBJbmMxGTAXBgNVBAsTEHd3dy5kaWdpY2VydC5j\n" \
+"b20xIDAeBgNVBAMTF0RpZ2lDZXJ0IEdsb2JhbCBSb290IEcyMIIBIjANBgkqhkiG\n" \
+"9w0BAQEFAAOCAQ8AMIIBCgKCAQEAuzfNNNx7a8myaJCtSnX/RrohCgiN9RlUyfuI\n" \
+"2/Ou8jqJkTx65qsGGmvPrC3oXgkkRLpimn7Wo6h+4FR1IAWsULecYxpsMNzaHxmx\n" \
+"1x7e/dfgy5SDN67sH0NO3Xss0r0upS/kqbitOtSZpLYl6ZtrAGCSYP9PIUkY92eQ\n" \
+"q2EGnI/yuum06ZIya7XzV+hdG82MHauVBJVJ8zUtluNJbd134/tJS7SsVQepj5Wz\n" \
+"tCO7TG1F8PapspUwtP1MVYwnSlcUfIKdzXOS0xZKBgyMUNGPHgm+F6HmIcr9g+UQ\n" \
+"vIOlCsRnKPZzFBQ9RnbDhxSJITRNrw9FDKZJobq7nMWxM4MphQIDAQABo0IwQDAP\n" \
+"BgNVHRMBAf8EBTADAQH/MA4GA1UdDwEB/wQEAwIBhjAdBgNVHQ4EFgQUTiJUIBiV\n" \
+"5uNu5g/6+rkS7QYXjzkwDQYJKoZIhvcNAQELBQADggEBAGBnKJRvDkhj6zHd6mcY\n" \
+"1Yl9PMWLSn/pvtsrF9+wX3N3KjITOYFnQoQj8kVnNeyIv/iPsGEMNKSuIEyExtv4\n" \
+"NeF22d+mQrvHRAiGfzZ0JFrabA0UWTW98kndth/Jsw1HKj2ZL7tcu7XUIOGZX1NG\n" \
+"Fdtom/DzMNU+MeKNhJ7jitralj41E6Vf8PlwUHBHQRFXGU7Aj64GxJUTFy8bJZ91\n" \
+"8rGOmaFvE7FBcf6IKshPECBV1/MUReXgRPTqh5Uykw7+U0b6LJ3/iyK5S9kJRaTe\n" \
+"pLiaWN0bfVKfjllDiIGknibVb63dDcY3fe0Dkhvld1927jyNxF1WW6LZZm6zNTfl\n" \
+"MrY=\n" \
+"-----END CERTIFICATE-----"
 
 #endif // APP_CONFIG_H
