+++ source/app_task.c
@@ -63,15 +63,11 @@
 #include "app_config.h"
 #include "app_task.h"
 
-
 #include "cy_log.h"
-
-/* OTA API */
 #include "cy_ota_api.h"
-#include "ota_serial_flash.h"
 
-/* App specific configuration */
-#include "ota_app_config.h"
+#include "ota_driver.h"
+#include "ota_serial_flash.h"
 
 #define APP_VERSION "01.00.00"
 
@@ -84,61 +80,14 @@
 * Forward declaration
 ********************************************************************************/
 
-cy_ota_callback_results_t ota_callback(cy_ota_cb_struct_t *cb_data);
-void print_heap_usage(char *msg);
 
 
 /*******************************************************************************
 * Global Variables
 ********************************************************************************/
 /* OTA context */
-cy_ota_context_ptr ota_context;
 volatile static bool otaFlag = false;
 
-/* Network parameters for OTA */
-cy_ota_network_params_t ota_network_params =
-{
-    .http =
-    {
-        .server =
-        {
-            .host_name = HTTP_SERVER,
-            .port = HTTP_SERVER_PORT
-        },
-        .file = OTA_HTTP_JOB_FILE,
-    #if (ENABLE_TLS == true)
-        .credentials =
-        {
-            .root_ca = ROOT_CA_CERTIFICATE,
-            .root_ca_size = sizeof(ROOT_CA_CERTIFICATE),
-        #if (USING_CLIENT_CERTIFICATE == true)
-            .client_cert = CLIENT_CERTIFICATE,
-            .client_cert_size = sizeof(CLIENT_CERTIFICATE),
-        #endif
-        #if (USING_CLIENT_KEY == true)
-            .private_key = CLIENT_KEY,
-            .private_key_size = sizeof(CLIENT_KEY),
-        #endif
-        },
-    #endif
-    },
-    .use_get_job_flow = CY_OTA_DIRECT_FLOW,
-#if (ENABLE_TLS == true)
-    .initial_connection = CY_OTA_CONNECTION_HTTPS,
-#else
-    .initial_connection = CY_OTA_CONNECTION_HTTP
-#endif
-};
-
-/* Parameters for OTA agent */
-cy_ota_agent_params_t ota_agent_params =
-{
-    .cb_func = ota_callback,
-    .cb_arg = &ota_context,
-    .reboot_upon_completion = 1,
-    .validate_after_reboot = 1,
-    .do_not_send_result = 1
-};
 
 /* Macro to check if the result of an operation was successful and set the
  * corresponding bit in the status_flag based on 'init_mask' parameter. When
@@ -161,196 +110,6 @@ cy_ota_agent_params_t ota_agent_params =
 
 
 
-/*******************************************************************************
- * Function Name: ota_callback()
- *******************************************************************************
- * Summary:
- *  Prints the status of the OTA agent on every event. This callback is optional,
- *  but be aware that the OTA middleware will not print the status of OTA agent
- *  on its own.
- *
- * Return:
- *  CY_OTA_CB_RSLT_OTA_CONTINUE - OTA Agent to continue with function.
- *  CY_OTA_CB_RSLT_OTA_STOP     - OTA Agent to End current update session.
- *  CY_OTA_CB_RSLT_APP_SUCCESS  - Application completed task, success.
- *  CY_OTA_CB_RSLT_APP_FAILED   - Application completed task, failure.
- *
- *******************************************************************************/
-cy_ota_callback_results_t ota_callback(cy_ota_cb_struct_t *cb_data)
-{
-    cy_ota_callback_results_t   cb_result = CY_OTA_CB_RSLT_OTA_CONTINUE;
-    const char                  *state_string;
-    const char                  *error_string;
-
-    if (cb_data == NULL)
-    {
-        return CY_OTA_CB_RSLT_OTA_STOP;
-    }
-
-    state_string  = cy_ota_get_state_string(cb_data->ota_agt_state);
-    error_string  = cy_ota_get_error_string(cy_ota_get_last_error());
-
-    print_heap_usage("In OTA Callback");
-
-    switch (cb_data->reason)
-    {
-
-        case CY_OTA_LAST_REASON:
-            break;
-
-        case CY_OTA_REASON_SUCCESS:
-            printf(">> APP CB OTA SUCCESS state:%d %s last_error:%s\n\n",
-                    cb_data->ota_agt_state,
-                    state_string, error_string);
-            break;
-
-        case CY_OTA_REASON_FAILURE:
-            printf(">> APP CB OTA FAILURE state:%d %s last_error:%s\n\n",
-                    cb_data->ota_agt_state, state_string, error_string);
-            break;
-
-        case CY_OTA_REASON_STATE_CHANGE:
-            switch (cb_data->ota_agt_state)
-            {
-                case CY_OTA_STATE_NOT_INITIALIZED:
-                case CY_OTA_STATE_EXITING:
-                case CY_OTA_STATE_INITIALIZING:
-                case CY_OTA_STATE_AGENT_STARTED:
-                case CY_OTA_STATE_AGENT_WAITING:
-                    break;
-
-                case CY_OTA_STATE_START_UPDATE:
-                    printf("APP CB OTA STATE CHANGE CY_OTA_STATE_START_UPDATE\n");
-                    break;
-
-                case CY_OTA_STATE_JOB_CONNECT:
-                    printf("APP CB OTA CONNECT FOR JOB using ");
-                    /* NOTE:
-                     *  HTTP - json_doc holds the HTTP "GET" request
-                     */
-                    if ((cb_data->broker_server.host_name == NULL) ||
-                        ( cb_data->broker_server.port == 0) ||
-                        ( strlen(cb_data->file) == 0) )
-                    {
-                        printf("ERROR in callback data: HTTP: server: %p port: %d topic: '%p'\n",
-                                cb_data->broker_server.host_name,
-                                cb_data->broker_server.port,
-                                cb_data->file);
-                        cb_result = CY_OTA_CB_RSLT_OTA_STOP;
-                    }
-                    printf("HTTP: server:%s port: %d file: '%s'\n",
-                            cb_data->broker_server.host_name,
-                            cb_data->broker_server.port,
-                            cb_data->file);
-
-                    break;
-
-                case CY_OTA_STATE_JOB_DOWNLOAD:
-                    printf("APP CB OTA JOB DOWNLOAD using ");
-                    /* NOTE:
-                     *  HTTP - json_doc holds the HTTP "GET" request
-                     */
-                    printf("HTTP: '%s'\n", cb_data->file);
-                    break;
-
-                case CY_OTA_STATE_JOB_DISCONNECT:
-                    printf("APP CB OTA JOB DISCONNECT\n");
-                    break;
-
-                case CY_OTA_STATE_JOB_PARSE:
-                    printf("APP CB OTA PARSE JOB: '%.*s' \n",
-                    strlen(cb_data->json_doc),
-                    cb_data->json_doc);
-                    break;
-
-                case CY_OTA_STATE_JOB_REDIRECT:
-                    printf("APP CB OTA JOB REDIRECT\n");
-                    break;
-
-                case CY_OTA_STATE_DATA_CONNECT:
-                    printf("APP CB OTA CONNECT FOR DATA using ");
-                    printf("HTTP: %s:%d \n", cb_data->broker_server.host_name,
-                    cb_data->broker_server.port);
-                    break;
-
-                case CY_OTA_STATE_DATA_DOWNLOAD:
-                    printf("APP CB OTA DATA DOWNLOAD using ");
-                    /* NOTE:
-                     *  HTTP - json_doc holds the HTTP "GET" request
-                     */
-                    printf("HTTP: '%.*s' ", strlen(cb_data->json_doc), cb_data->json_doc);
-                    printf("File: '%s'\n\n", cb_data->file);
-                    break;
-
-                case CY_OTA_STATE_DATA_DISCONNECT:
-                    printf("APP CB OTA DATA DISCONNECT\n");
-                    break;
-
-                case CY_OTA_STATE_RESULT_CONNECT:
-                    printf("APP CB OTA SEND RESULT CONNECT using ");
-                    /* NOTE:
-                     *  HTTP - json_doc holds the HTTP "GET" request
-                     */
-                    printf("HTTP: Server:%s port: %d\n",
-                            cb_data->broker_server.host_name,
-                            cb_data->broker_server.port);
-                    break;
-
-                case CY_OTA_STATE_RESULT_SEND:
-                    printf("APP CB OTA SENDING RESULT using ");
-                    /* NOTE:
-                     *  HTTP - json_doc holds the HTTP "PUT"
-                     */
-                    printf("HTTP: '%s' \n", cb_data->json_doc);
-                    break;
-
-                case CY_OTA_STATE_RESULT_RESPONSE:
-                    printf("APP CB OTA Got Result response\n");
-                    break;
-
-                case CY_OTA_STATE_RESULT_DISCONNECT:
-                    printf("APP CB OTA Result Disconnect\n");
-                    break;
-
-                case CY_OTA_STATE_OTA_COMPLETE:
-                    printf("APP CB OTA Session Complete\n");
-                    break;
-
-                case CY_OTA_STATE_STORAGE_OPEN:
-                    printf("APP CB OTA STORAGE OPEN\n");
-                    break;
-
-                case CY_OTA_STATE_STORAGE_WRITE:
-                    printf("APP CB OTA STORAGE WRITE %ld%% (%ld of %ld)\n",
-                            (unsigned long)cb_data->percentage,
-                            (unsigned long)cb_data->bytes_written,
-                            (unsigned long)cb_data->total_size);
-
-                    /* Move cursor to previous line */
-                    printf("\x1b[1F");
-                    break;
-
-                case CY_OTA_STATE_STORAGE_CLOSE:
-                    printf("APP CB OTA STORAGE CLOSE\n");
-                    break;
-
-                case CY_OTA_STATE_VERIFY:
-                    printf("APP CB OTA VERIFY\n");
-                    break;
-
-                case CY_OTA_STATE_RESULT_REDIRECT:
-                    printf("APP CB OTA RESULT REDIRECT\n");
-                    break;
-
-                case CY_OTA_NUM_STATES:
-                    break;
-            }   /* switch state */
-            break;
-    }
-
-    return cb_result;
-}
-
 
 static void publish_telemetry() {
     IotclMessageHandle msg = iotcl_telemetry_create();
@@ -368,52 +127,8 @@ static void publish_telemetry() {
     iotcl_destroy_serialized(str);
 }
 
-static bool spliturl(const char *url, char **host_name, char**resource) {
-    int host_name_start = 0;
-    size_t url_len = strlen(url);
 
-    if (!host_name || !resource) {
-        printf("split_url: Invalid usage\r\n");
-        return false;
-    }
-
-    *host_name = NULL;
-    *resource = NULL;
-    int slash_count = 0;
-    for (size_t i = 0; i < url_len; i++) {
-        if (url[i] == '/') {
-            slash_count++;
-            if (slash_count == 2) {
-                host_name_start = i + 1;
-            } else if (slash_count == 3) {
-                const size_t slash_start = i;
-                const size_t host_name_len = i - host_name_start;
-                const size_t resource_len = url_len - i;
-                *host_name = malloc(host_name_len + 1); //+1 for null
-                if (NULL == *host_name) {
-                    return false;
-                }
-                memcpy(*host_name, &url[host_name_start], host_name_len);
-                (*host_name)[host_name_len] = 0; // terminate the string
-
-                *resource = malloc(resource_len + 1); //+1 for null
-                if (NULL == *resource) {
-                    free(*host_name);
-                    return false;
-                }
-                memcpy(*resource, &url[slash_start], resource_len);
-                (*resource)[resource_len] = 0; // terminate the string
-
-                return true;
-            }
-        }
-    }
-    return false; // URL could not be parsed
-}
-
-
-
-void start_ota(IotclEventData data)
+void on_ota(IotclEventData data)
 {
 	otaFlag = true;
 
@@ -423,23 +138,24 @@ void start_ota(IotclEventData data)
 	char **path = &otapath;
 
     char *url = iotcl_clone_download_url(data, 0);
+    if (url == NULL){
+    	printf("Download URL is invalid.\r\n");
+    	return;
+    }
 //    printf("\r\nURL is %s\r\n", url);
 
     bool status = spliturl(url, host, path);
     if (!status) {
         printf("start_ota: Error while splitting the URL, code: 0x%x\r\n", status);
     }
-    printf("\r\nHOST is %s\r\nPATH is %s\r\n", otahost, otapath);
 
-    ota_network_params.http.file = otapath;
-    printf("\r\nHTTP FILE is %s\r\n", ota_network_params.http.file);
+//    ota_network_params.http.file = otapath;
+//    ota_network_params.http.server.host_name = otahost;
 
-//    free(host);
-//    free(path);
-    if( cy_ota_agent_start(&ota_network_params, &ota_agent_params, &ota_context) != CY_RSLT_SUCCESS )
-    {
-        printf("\n Initializing and starting the OTA agent failed.\n");
-        CY_ASSERT(0);
+    printf("\r\nHOST is %s.\r\nPATH is %s.\r\n", otahost, otapath);
+
+    if(start_ota(otahost, otapath)){
+    	printf("OTA starts successfully.\r\n");
     }
 }
 
@@ -535,7 +251,7 @@ void app_task(void *pvParameters) {
             iotc_config->auth.data.cert_info.device_cert = IOTCONNECT_DEVICE_CERT;
             iotc_config->auth.data.cert_info.device_key = IOTCONNECT_DEVICE_KEY;
         }
-        iotc_config->ota_cb = start_ota;
+        iotc_config->ota_cb = on_ota;
 
         cy_rslt_t ret = iotconnect_sdk_init();
         if (CY_RSLT_SUCCESS != ret) {
